dictmaker/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
